+++ src/app/dashboard/layout.tsx
@@ -1,4 +1,4 @@
-import { Sidebar } from "../components";
+import { Sidebar } from "../../components/Sidebar";
 
 export default function DashboardLayout({children}: {children: React.ReactNode;}) {
   return (
