+++ src/app/dashboard/pokemons/page.tsx
@@ -1,8 +1,10 @@
-import { PokemonGrid, PokemonsResponse, SimplePokemon } from "@/pokemons";
+import { PokemonGrid } from "@/pokemons/components/PokemonGrid";
+import { PokemonsResponse } from "@/pokemons/interfaces/pokemons-response";
+import { SimplePokemon } from "@/pokemons/interfaces/simple-pokemon";
 
 
 const getPokemons = async ( limit = 20, offset = 0 ):Promise<SimplePokemon[]> => {
-    const data:PokemonsResponse = await fetch(`https://pokeapi.co/api/v2/pokemon?limit=${limit}&offset=${offset}}`)
+    const data:PokemonsResponse = await fetch(`https://pokeapi.co/api/v2/pokemon?limit=${limit}&offset=${offset}`)
     .then( res => res.json());
 
     const pokemons = data.results.map( pokemon => ({
@@ -10,8 +12,7 @@ const getPokemons = async ( limit = 20, offset = 0 ):Promise<SimplePokemon[]> =>
         name: pokemon.name,
     }) );
 
-    //throw new Error('Esto es un error que no debria suceder')
-
+  
     return pokemons;
 }
